Extract TextField and rebind error & label
