+++ src/atomic/index.js
@@ -1,2 +1,3 @@
 export * from './atoms/MUI-Wrappers';
 export { default as SelectField } from './atoms/SelectField';
+export { default as TextField } from './atoms/TextField';
